fix: rename the file services correctly
