+++ src/components/partner-discount/PriceTable.tsx
@@ -2,10 +2,10 @@
 import React, { useEffect, useState } from 'react';
 import { Transition } from '@headlessui/react';
 import PriceCard from '@/components/partner-discount/PriceCard';
-import { ProductsProps, stripeService } from '@/components/services/stripeService';
+import { ProductsProps, stripeService } from '@/components/services/stripe.service';
 import CardSkeleton from '@/components/components/CardSkeleton';
 import { CouponType } from '@/pages/api/stripe/get_coupons';
-import { currencyService } from '@/components/services/currencyService';
+import { currencyService } from '@/components/services/currency.service';
 
 export default function PriceTable({ lang }: { lang: string }) {
   const [products, setProducts] = useState<ProductsProps>();
