+++ src/components/partnerships/cloudwards/components/ButtonDeal.tsx
@@ -2,8 +2,8 @@ import React, { useEffect } from 'react';
 
 import { checkout } from '../../../../lib/auth';
 import { CouponType } from '../../../../pages/api/stripe/get_coupons';
-import { Interval, stripeService } from '../../../services/stripeService';
-import { currencyService } from '../../../services/currencyService';
+import { Interval, stripeService } from '../../../services/stripe.service';
+import { currencyService } from '../../../services/currency.service';
 
 const ButtonDeal = ({ textContent, large }) => {
   const [priceId, setPriceId] = React.useState('');
