+++ src/components/prices/CyberAwarenessSpecialCard.tsx
@@ -7,8 +7,8 @@
 import React, { useEffect } from 'react';
 import { checkout } from '@/lib/auth';
 import { CouponType } from '@/pages/api/stripe/get_coupons';
-import { analyticsService } from '@/components/services/analyticsService';
-import { stripeService } from '@/components/services/stripeService';
+import { analyticsService } from '@/components/services/analytics.service';
+import { stripeService } from '@/components/services/stripe.service';
 import { PriceCardProps } from './PriceCard';
 
 export default function SpecialPriceCard({
