+++ src/components/prices/PriceTable.tsx
@@ -4,9 +4,9 @@ import { Transition } from '@headlessui/react';
 import PriceCard from './PriceCard';
 import { Coin, CreditCard, Detective } from '@phosphor-icons/react';
 import BusinessBanner from '@/components/banners/BusinessBanner';
-import { Interval, ProductsProps, stripeService } from '@/components/services/stripeService';
+import { Interval, ProductsProps, stripeService } from '@/components/services/stripe.service';
 import CardSkeleton from '@/components/components/CardSkeleton';
-import { currencyService } from '@/components/services/currencyService';
+import { currencyService } from '@/components/services/currency.service';
 import FreePlanCard from './FreePlanCard';
 import Header from '@/components/shared/Header';
 import useStripeAndCurrency from '@/hooks/useProducts';
