+++ src/components/prices/SpecialPriceCard.tsx
@@ -8,8 +8,8 @@
 import React, { useEffect } from 'react';
 import { checkout } from '@/lib/auth';
 import { CouponType } from '@/pages/api/stripe/get_coupons';
-import { analyticsService } from '@/components/services/analyticsService';
-import { Interval, stripeService } from '@/components/services/stripeService';
+import { analyticsService } from '@/components/services/analytics.service';
+import { Interval, stripeService } from '@/components/services/stripe.service';
 import { PriceCardProps } from './PriceCard';
 
 export default function SpecialPriceCard({
