+++ src/components/services/currency.service.ts
