+++ src/hooks/useProducts.ts
@@ -1,5 +1,5 @@
-import { currencyService } from '@/components/services/currencyService';
-import { ProductsProps, stripeService } from '@/components/services/stripeService';
+import { currencyService } from '@/components/services/currency.service';
+import { ProductsProps, stripeService } from '@/components/services/stripe.service';
 import { useEffect, useState } from 'react';
 
 function useStripeAndCurrency() {
